refactor(People): Names에서 People로 이름 변경
